Add downloadGtoFileByURI() method.
Split the code which downloads a single gto file given a uri out of the
downloadGtoFilesByURI() into the new method and just have
downloadGtoFilesByURI() loop over the uris, calling the new method each
iteration.

This is prepatory work to allow downloading a single gto file and then
immediately downloading the torrent data (BAM file). The
downloadGtoFilesByURI() method will likely be removed in the future.



git-svn-id: svn+ssh://svn/srv/svnroot/annaisystems-genetorrent/newGeneTorrent/trunk@901 ce0a4a09-f1da-4806-a783-9435acac145d