libtorrent: make libtorrent::session constructor more exception-safe
Before this change, if an attempt to create the network thread in libtorrent
failed during session construction, the session_impl destructor would hang
when it tried to join the disk io thread (which was already created).

The way the session_impl destructor is set up, it relies on the network
thread to do some abort tasks which cause the disk io thread to exit.

Now, if the disk io thread is successfully created but the attempt
to create the network thread fails, there's a destructor path that
will cause the session_impl to be torn down and the exception
will propagate to the caller of libtorrent::session().

In the normal, the network thread will exist and this
change will have no effect.

git-svn-id: svn+ssh://svn/srv/svnroot/annaisystems-genetorrent/newGeneTorrent/trunk@933 ce0a4a09-f1da-4806-a783-9435acac145d