CGHUBDEV-36: Change curl options to verify SSL certs and hosts
Set CURLOPT_SSL_VERIFYPEER and CURLOPT_SSL_VERIFYHOST to true
so that peer certificates are verified and the host identity matches
the host the certificate was issued for.

Add command line and configuration file option --ssl-no-verify-ca to toggle
whether libcurl verifies SSL certs and hosts.

Default is now true, so certs and hosts are verified.

Update man page to document option.

git-svn-id: svn+ssh://svn/srv/svnroot/annaisystems-genetorrent/newGeneTorrent/trunk@945 ce0a4a09-f1da-4806-a783-9435acac145d