CGHUBDEV-98: Modify disable_hash_checks behaviour.
Prior to this change, when the session disable_hash_checks flag was set,
the hash calculations where not performed.

In order to test download peer performance, this change makes the flag
cause the hashes to still be calculated, but just doesn't fail things if
the hashes do not match the expected values.

git-svn-id: svn+ssh://svn/srv/svnroot/annaisystems-genetorrent/newGeneTorrent/trunk@1003 ce0a4a09-f1da-4806-a783-9435acac145d