gtoinfo: emit_gto fails if file does not already exist
emit_gto is documented to work whether or not the destination file
already exists.  The os.stat() call throws an exception if
the destination file does not exist.

Fix emit_gto() so that it writes to the destination file if it
does not already exist.  Also, don't fail if a file backup
operation fails because the original file does not exist.

git-svn-id: svn+ssh://svn/srv/svnroot/annaisystems-genetorrent/newGeneTorrent/trunk@1017 ce0a4a09-f1da-4806-a783-9435acac145d