gtoinfo: bencode correctly on 32-bit platforms
In python, ints become longs when they exceed sys.maxint.  The int
type has the precision of a C int.  So, we need to bencode longs
as int types (bencode has no notion of longs).  Reading these back
(decoding) is ok because the type conversion to a long is automatic.
Longs support arbitrarily large numbers.

git-svn-id: svn+ssh://svn/srv/svnroot/annaisystems-genetorrent/newGeneTorrent/trunk@1042 ce0a4a09-f1da-4806-a783-9435acac145d