GeneTorrent Client waits several minutes after transfer is complete... (Jira: GTXFER-59)
The peer connection management code in libtorrent has a race condition where
SSL peers are not fully disconnected under certain circumstances.  This
has only been observed when a download client is attempting to disconnect
from the server after it is complete.  When the race works out such
that the socket is not closed when expected, the shutdown sequence in the
client is hung until the socket is eventually closed.  The time of the
eventual closure is dictated by the server's peer timeout value.  When
half of the peer timeout has expired, the sever sends a BT keepalive message
which finally triggers the socket closure for the hung peer connection in
the download client.

To address this, two changes in libtorrent are made.  The first, in
peer_connection.cpp, special-cases the SSL shutdown sequence in these
circumstance where the reason for disconnecting is "upload to upload
connection," which covers the above-mentioned case where the download client
has completed its download and is attempting to disconnect from the
server.  Instead of allowing a boost asio asynchronous SSL shutdown,
which may never complete, it attempts a SSL shutdown and then forcibly
closes the socket even if the shutdown fails.  This is no worse than
the previous behavior, where the SSL session shutdown is also not clean.

In my testing, this resolves the session destructor hang in libtorrent
which causes the shutdown delay described in the Jira ticket.

To catch any other code paths that may lead to the same hang, the second
change to libtorrent explicitly stops the networking thread IO service
in the libtorrent session destructor.  This forces libtorrent's
destructor to return even if there are outstanding connections.