GeneTorrent Client waits several minutes after transfer is complete... (Jira: GTXFER-59)
Follow up to commit 0f16176478

It is not safe to stop the IO service in the session destructor because
it may have outstanding operations that must complete before other destructors
are called.

For example, if there's an outstanding UDP operation, it must be completed
before the udp_socket is destructed because the udp_socket object's
destructor asserts that oustanding ops is 0.

This was a failsafe in the previous commit.  Other changes in that commit
should have resolved GTXFER-59 without this change.