Download GTO file to a tmpfile.
When making the curl call to download a GTO file, download it to a
GTO.tmp file and only rename it to the GTO file after the downloaded
file has been processed for errors. If  the downloaded file contains an
error message embedded in XML, the error processor will terminate the
downloader before renaming the file.

This eliminates the potential to have a GTO file that contains XML
instead of valid torrent information.