GTXFER-65: Disconnect peer if connect stalls.
When the server tries to make an outgoing connection to the client peer,
the connection may not succeed due to firewalling or blocked ports. This
blocked connection will never succeed, but it is blocking (until it
times out) valid incoming connections from the client peer which slows
down the creation of the peer connection.

The fix is to disconnect the peer that is stalled instead of the newer
incoming connection (which is likely to succeed since it is incoming
instead of outgoing).