This commit introduces downloading GTO data to a temporary location.
A fundamental change in the behavior of GeneTorrent download clients is delivered with this commit.

If the final destination UUID is present in the directory where gtdownload is executing,
gtdownload's behavior is to treat the UUID directory as a full and completely downloaded GTO.
This behavior occurs regardless of the type of download processed (e.g., from a UUID, from an XML
input, an existing .gto file, or URL).

Signed-off-by: donavan nelson <donavan@cardinalpeak.com>