Remove -Werror -Wextra from CXXFLAGS
With -Werror, it was just too hard to maintain the warnings across different platforms and compiler versions. Newer compilers get more picky and required additional warning suppressions  via -Wno-... switches.

Additionally, the

#pragma GCC diagnostic ignored "-Wfoo"
#include ...
#pragma GCC diagnostic error "-Wfoo"

stuff was getting out of control. It's just too brittle because it assumes that the foo warning is enabled via a command line flag. If it isn't, the second pragma will turn it on, even if -Wno-foo was specified on the command line.

There is a benefit to -Werror and -Wextra but the hassle maintaining it across platforms outweighs it.