Fix: gtdownload on Mac OS X fails to load with "image not found" for libxerces-c-3.1.dylib
The exact error is

dyld: Library not loaded: /Users/jenkins/workspace/genetorrent-osx/deps/xerces-c/lib/libxerces-c-3.1.dylib
  Referenced from: /usr/local/lib/GeneTorrent/libxqilla.6.dylib
  Reason: image not found
Trace/BPT trap: 5

This was caused by an extraneous space in the first argument to install_name_tool (the dylib entry to be changed). Because we originally override IFS in this recipe to just newline, bash did not discard the space like it usually does when interpolating a variable. Due to the extra space, install_name_tool did not match the dyld name and therefore refused to make the change. It did return 0 still which is why we didn't catch it. Did I mention that I hate make and shell programming? It's so arcane.

Adding space to IFS should fix this.

This is a regression introduced by 4b3d9c5b58947d5a5e45e1d165c77688fe0d130f, which is still needed as a workaround for the crazy bug mention there.

We should consider adding space to all other IFS overrides in Makefile.am. Due to us being close to release I am postponing that for now, though.